+++ module/effects-panel.mjs
@@ -10,6 +10,18 @@ export class EffectsPanel extends Application {
         };
     }
 
+    // Addlistners
+    // effect-item
+    activateListeners(html) {
+        super.activateListeners(html);
+        html.find(".effect-item").click(this._onEffectClick.bind(this));
+        html.find(".effect-item").mouseover(this._onEffectClick.bind(this));
+    }
+
+    _onEffectClick(event) {
+        console.log(event);
+    }
+
     get token() {
         return canvas.tokens.controlled.at(0)?.document ?? null;
     }
@@ -28,6 +40,20 @@ export class EffectsPanel extends Application {
     async getData(options) {
         const context = await super.getData(options);
         context.effects = this.actor?.appliedEffects; //effects;
+
+        if (context.effects) {
+            for (const ae of context.effects) {
+                // Sometimes ae.parent?.system.active  is false, but the power is active, unclear why.
+                // Consider making active a getter (looking for the AE) instead of using system.actor.
+                if (ae.parent instanceof HeroSystem6eItem && ae.duration.seconds) {
+                    ae.flags.label = `${ae.flags.startTime + ae.duration.seconds - game.time.worldTime} seconds`;
+                } else {
+                    const d = ae._prepareDuration();
+                    ae.flags.label = d.label;
+                }
+            }
+        }
+
         return context;
     }
 }