+++ module/item/item.mjs
@@ -444,7 +444,7 @@ export class HeroSystem6eItem extends Item {
         let item = this;
 
         if (!item.system.active) {
-            if (!this.actor.canAct(true)) {
+            if (!this.actor.canAct(true, event)) {
                 return;
             }
 
@@ -452,13 +452,118 @@ export class HeroSystem6eItem extends Item {
             let end = parseInt(this.system.end);
             let value = parseInt(this.actor.system.characteristics.end.value);
             if (end > value) {
-                if (event?.ctrlKey) {
-                    ui.notifications.info(`${game.user.name} used CTRL key to force <b>${this.name}</b> on.`);
+                if (event?.shiftKey) {
+                    const speaker = ChatMessage.getSpeaker({
+                        actor: this,
+                        //token,
+                    });
+                    speaker["alias"] = game.user.name;
+                    const chatData = {
+                        user: game.user._id,
+                        type: CONST.CHAT_MESSAGE_TYPES.OTHER,
+                        content: `${game.user.name} used SHIFT key to force <b>${this.name}</b> on.`,
+                        whisper: whisperUserTargetsForActor(this),
+                        speaker,
+                    };
+                    ChatMessage.create(chatData);
                 } else {
                     ui.notifications.error(
-                        `Unable to active ${this.name}.  ${item.actor.name} has ${value} END.  Power requires ${end} END to activate.  Hold CTRL to force.`,
+                        `Unable to active ${this.name}.  ${item.actor.name} has ${value} END.  Power requires ${end} END to activate.  Hold SHIFT to force.`,
                     );
                     return;
+                }
+            }
+
+            // Spend CHARGE to toggle power on
+            // Notice that item.system.charges is used to provide details
+            const charges = this.findModsByXmlid("CHARGES");
+            if (charges) {
+                if (this.system.charges.value <= 0) {
+                    if (event?.shiftKey) {
+                        const speaker = ChatMessage.getSpeaker({
+                            actor: this,
+                            //token,
+                        });
+                        speaker["alias"] = game.user.name;
+                        const chatData = {
+                            user: game.user._id,
+                            type: CONST.CHAT_MESSAGE_TYPES.OTHER,
+                            content: `${game.user.name} used SHIFT key to force <b>${this.name}</b> on.`,
+                            whisper: whisperUserTargetsForActor(this),
+                            speaker,
+                        };
+                        ChatMessage.create(chatData);
+                    } else {
+                        ui.notifications.error(
+                            `Unable to active ${this.name}.  ${item.actor.name} has ${this.system.charges.value} CHARGES.  Hold SHIFT to force.`,
+                        );
+                        return;
+                    }
+                }
+                this.system.charges.value -= 1;
+                this.update({ "system.charges": this.system.charges });
+
+                // Charges expire, find the Active Effect
+                const ae = this.effects.contents?.[0];
+                if (ae) {
+                    let seconds = 1;
+                    const continuing = this.findModsByXmlid("CONTINUING");
+                    if (continuing) {
+                        // TODO: Extract (look in adjustment)
+                        switch (continuing.OPTIONID) {
+                            case "EXTRAPHASE":
+                                seconds = 2;
+                                break;
+                            case "TURN":
+                                seconds = 12;
+                                break;
+                            case "MINUTE":
+                                seconds = 60;
+                                break;
+                            case "FIVEMINUTES":
+                                seconds = 60 * 5;
+                                break;
+                            case "TWENTYMINUTES":
+                                seconds = 60 * 20;
+                                break;
+                            case "HOUR":
+                                seconds = 60 * 60;
+                                break;
+                            case "SIXHOURS":
+                                seconds = 60 * 60 * 6;
+                                break;
+                            case "ONEDAY":
+                                seconds = 60 * 60 * 24;
+                                break;
+                            case "ONEWEEK":
+                                seconds = 60 * 60 * 24 * 7;
+                                break;
+                            case "ONEMONTH":
+                                seconds = 60 * 60 * 24 * 30;
+                                break;
+                            case "ONESEASON":
+                                seconds = 60 * 60 * 24 * 90;
+                                break;
+                            case "ONEYEAR":
+                                seconds = 60 * 60 * 24 * 365;
+                                break;
+                            case "FIVEYEARS":
+                                seconds = 60 * 60 * 24 * 365 * 5;
+                                break;
+                            case "TWENTYFIVEYEARS":
+                                seconds = 60 * 60 * 24 * 365 * 25;
+                                break;
+                            case "ONECENTURY":
+                                seconds = 60 * 60 * 24 * 365 * 100;
+                                break;
+                        }
+                    }
+
+                    console.log(
+                        await ae.update({ "duration.seconds": seconds, "flags.startTime": game.time.worldTime }),
+                    );
+                } else {
+                    console.log("No associated Active Effect", this);
                 }
             }
 
@@ -575,7 +680,12 @@ export class HeroSystem6eItem extends Item {
                 break;
         }
 
-        // If we have control of this token, reaquire to update movement types
+        // Charges expire
+        // if (charges) {
+        //     // Find the active effect
+        // }
+
+        // If we have control of this token, re-acquire to update movement types
         const myToken = this.actor?.getActiveTokens()?.[0];
         if (canvas.tokens.controlled.find((t) => t.id == myToken.id)) {
             myToken.release();
@@ -4475,12 +4585,16 @@ export class HeroSystem6eItem extends Item {
                 });
             }
 
-            for (const enhancedPerception of this.actor.items.filter((o) => o.system.XMLID === "ENHANCEDPERCEPTION")) {
-                if (enhancedPerception.system.checked && enhancedPerception.system.active) {
+            for (const enhancedPerception of this.actor.items.filter(
+                (o) => o.system.XMLID === "ENHANCEDPERCEPTION" && o.system.OPTIONID === "ALL",
+            )) {
+                enhancedPerception.system.checked = true;
+                if (enhancedPerception.system.active) {
                     const levels = parseInt(enhancedPerception.system.LEVELS);
                     tags.push({
                         value: levels,
                         name: enhancedPerception.name,
+                        itemId: enhancedPerception.id,
                     });
                     rollVal += levels;
                 }